// apps/exhibits_service/src/beetroot-services/dto/exhibitQuestion.dto.ts
import { Editor } from '../entities/beetroot-service.entity';
import { Questionnaire } from '../entities/questionaire.entity';

// DTO for ExhibitQuestionaire
export class Questionnairedto {
  questionnaireSectionID?: string;
  questionnaireID?:string;
  category?: string;
  searchTerms?: string;
  title?: string;
  body?: string;
  editor?: Editor; // Assuming you have a User DTO
  questions?: [];
}
export class QuestionnaireSectionDto {
  questionnaireID?: string;
  questionnaireSectionID?: string;
  title?: string;
  questions?: [];
}

export class ExhibitSocketPayloadDTO {
  questionnaireID:string;
  clientAuth: string;
  status: string;
  payload: string
  bookingType: string;
  questionnaireSectionID: string;
  clientID: string;
}
export class ExhibitSectionDto {
  exhibitID?: string;
  title?: string;
  questions?: [];
}
// DTO for ExhibitQuestionaire
export class ExhibitQuestionairedto {
  authToken?: string;
  questionnaire?: Questionnaire;
}

// DTO for ExhibitQuestion
export class QuestionDto {
  category: string;
  question: string;
  meta?: string;
  searchTerms?: string;
  title?: string;
  body?: string;
  editorId?: string;
  publisher: Editor; // Assuming you have a User DTO
  questionnaireIds: string[];
}